BEMOSS_requirements.txt
blist==1.3.6
decorator==4.0.9
dnspython==1.12.0
django==1.10
django-filter
djangorestframework==3.5.4
futures==3.0.5
lxml==3.2.3
netifaces==0.10.4
networkx==1.11
nose==1.3.3
Pillow==2.9.0
pandas==0.18.1
psycopg2==2.6.2
pycrypto==2.6.1
#pyOpenSSL==0.13.1
scales==1.0.9
sleekxmpp==1.3.1
spyrk==0.0.3
tornado==4.4.1
tablib==0.11.2
wheel==0.24.0

